rn_price_fetch/pyproject.toml
[tool.poetry]
name = "amazon-price-fetch"
version = "0.1.0"
description = "A robust Amazon price search and data extraction tool"
packages = [
    { include = "amazon_price_fetch", from = "src" },
]
authors = ["Ram Nath Bamal"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.7.4"
click = "^8.1.7"
pandas = "^2.2.2"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.2"
pyyaml = "^6.0.1"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
mypy = "^1.8.0"
black = "^24.1.0"
isort = "^5.13.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = "src"
ignore_missing_imports = true
follow_imports = "silent"
show_error_codes = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true
disallow_incomplete_defs = true

[tool.isort]
py_version = 311
combine_as_imports = true
profile = "black"
lines_between_types = 1
lines_after_imports = 2

[tool.black]
target-version = ['py311']
line-length = 100
skip-string-normalization = true